Adding in a 404 page :D (#77)
Off topic, but this PR removes artefacts being displayed as well as
adding in a wee 404 page onto our system.

The removal of artefacts and the addition of a 404 page should 100% have
been on separate PRs but I forgot that I had done it.

FYI !!!! 500 lines of the 700 are actually package lock json so it isn't
as bad as it looks